More testing changes
