+++ tests/Unit/Http/IncomingRequestTest.php
@@ -7,26 +7,8 @@ use PHPUnit\Framework\TestCase;
 require_once 'src/ShinePHP/Http/IncomingRequest.php';
 use ShinePHP\Http\IncomingRequest;
 
-final class IncomingRequestTest extends TestCase {
+final class IncomingRequestUnitTest extends TestCase {
 
-	public function testValidCreate() : void {
-
-		$Incoming = new IncomingRequest();
-
-		$this->assertInstanceOf(
-			IncomingRequest::class,
-			$Incoming
-		);
-
-	}
-
-	// Testing valid JSON input from url (will usually be from php://input though)
-	public function testingValidJsonInputFromUrl() : void {
-
-		$Incoming = new IncomingRequest();
-
-		$jsonRetrieved = $Incoming::retrieve_json_input('https://jsonplaceholder.typicode.com/posts');
-		$this->assertArrayHasKey(50, $jsonRetrieved);
-	}
+	// test the input sanitizer
 
 }
